+++ seed/factories/campaign.factory.ts
@@ -4,7 +4,6 @@ import { STATE_CODES } from '../../src/shared/constants/states'
 import { LEVELS } from '../../src/shared/constants/governmentLevels'
 import { generateFactory } from './generate'
 import { GenerationStatus } from 'src/campaigns/ai/content/aiContent.types'
-import { STATES } from 'api/dist/src/shared/constants/states'
 
 export const campaignFactory = generateFactory<Campaign>(() => {
   const electionDate = faker.date.past()
@@ -27,7 +26,7 @@ export const campaignFactory = generateFactory<Campaign>(() => {
       },
     },
     details: {
-      state: faker.helpers.arrayElement(STATES),
+      state: faker.helpers.arrayElement(STATE_CODES),
       zip: faker.location.zipCode(),
       ballotLevel: faker.helpers.arrayElement(LEVELS),
       electionDate: electionDate.toISOString().split('T')[0],
