+++ src/campaigns/campaigns.module.ts
@@ -2,8 +2,8 @@ import { forwardRef, Module } from '@nestjs/common'
 import { CampaignsController } from './campaigns.controller'
 import { CampaignsService } from './services/campaigns.service'
 import { CampaignsAiModule } from './ai/campaignsAi.module'
-import { CampaignMapController } from './mapping/campaignMap.controller'
-import { CampaignMapService } from './mapping/campaignMap.service'
+import { CampaignMapController } from './map/campaignMap.controller'
+import { CampaignMapService } from './map/campaignMap.service'
 import { CampaignPlanVersionsService } from './services/campaignPlanVersions.service'
 import { EmailModule } from 'src/email/email.module'
 import { CampaignPositionsController } from './positions/campaignPositions.controller'
