+++ src/email/mailgun.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common'
-import * as FormData from 'form-data'
+import FormData from 'form-data'
 import Mailgun, { MailgunMessageData } from 'mailgun.js'
 import { IMailgunClient } from 'mailgun.js/Interfaces'
